Update overview spec of `laneselect` instructions
Currently the overview indicates that if the mask lane is not zero,
not all-ones, and the top bit is 0 then the output must be the `b`
element.  This doesn't match (what I believe is) the intended
implementation on AArch64 using the `bsl` instruction which is the same
as `v128.bitselect`.